requirements.txt
absl-py==2.1.0
alpaca-py==0.38.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
dill==0.3.9
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
msgpack==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
numba==0.61.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandarallel==1.6.5
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyzmq==26.2.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sseclient-py==1.8.0
stack-data==0.6.3
ta==0.11.0
termcolor==2.5.0
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
websockets==15.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==2.1.4
